Rename congestion_ctrl module (and directory) to congestion_control.
